css 수정 + member랑 planner vo에 warningCount 필드 추가

+++ tripRecord/src/main/java/com/finalproject/triprecord/matching/model/vo/ReqSchedule.java
@@ -18,7 +18,7 @@ public class ReqSchedule {
 	private int scheNo;
 	private String reqRef;
 	private int reqStatus;
-	private double payPoint;
+	private int payPoint;
 	private Date confirmDate;
 	private String nickname;
